+++ src/Aws/AutoScaling/Resources/autoscaling-2011-01-01.php
@@ -112,6 +112,51 @@ return array (
                 ),
             ),
         ),
+        'CompleteLifecycleAction' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'CompleteLifecycleAction',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2011-01-01',
+                ),
+                'LifecycleHookName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 255,
+                ),
+                'AutoScalingGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+                'LifecycleActionToken' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 36,
+                    'maxLength' => 36,
+                ),
+                'LifecycleActionResult' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                ),
+            ),
+        ),
         'CreateAutoScalingGroup' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -578,6 +623,39 @@ return array (
                 ),
             ),
         ),
+        'DeleteLifecycleHook' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'DeleteLifecycleHook',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2011-01-01',
+                ),
+                'LifecycleHookName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 255,
+                ),
+                'AutoScalingGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+            ),
+        ),
         'DeleteNotificationConfiguration' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -915,6 +993,62 @@ return array (
                 ),
             ),
         ),
+        'DescribeLifecycleHookTypes' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'DescribeLifecycleHookTypesAnswer',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'DescribeLifecycleHookTypes',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2011-01-01',
+                ),
+            ),
+        ),
+        'DescribeLifecycleHooks' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'DescribeLifecycleHooksAnswer',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'DescribeLifecycleHooks',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2011-01-01',
+                ),
+                'AutoScalingGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+                'LifecycleHookNames' => array(
+                    'type' => 'array',
+                    'location' => 'aws.query',
+                    'sentAs' => 'LifecycleHookNames.member',
+                    'items' => array(
+                        'name' => 'AsciiStringMaxLen255',
+                        'type' => 'string',
+                        'minLength' => 1,
+                        'maxLength' => 255,
+                    ),
+                ),
+            ),
+        ),
         'DescribeMetricCollectionTypes' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -1237,6 +1371,49 @@ return array (
                 ),
             ),
         ),
+        'DetachInstances' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'DetachInstancesAnswer',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'DetachInstances',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2011-01-01',
+                ),
+                'InstanceIds' => array(
+                    'type' => 'array',
+                    'location' => 'aws.query',
+                    'sentAs' => 'InstanceIds.member',
+                    'items' => array(
+                        'name' => 'XmlStringMaxLen16',
+                        'type' => 'string',
+                        'minLength' => 1,
+                        'maxLength' => 16,
+                    ),
+                ),
+                'AutoScalingGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+                'ShouldDecrementDesiredCapacity' => array(
+                    'required' => true,
+                    'type' => 'boolean',
+                    'format' => 'boolean-string',
+                    'location' => 'aws.query',
+                ),
+            ),
+        ),
         'DisableMetricsCollection' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -1318,6 +1495,49 @@ return array (
                 ),
             ),
         ),
+        'EnterStandby' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'EnterStandbyAnswer',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'EnterStandby',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2011-01-01',
+                ),
+                'InstanceIds' => array(
+                    'type' => 'array',
+                    'location' => 'aws.query',
+                    'sentAs' => 'InstanceIds.member',
+                    'items' => array(
+                        'name' => 'XmlStringMaxLen16',
+                        'type' => 'string',
+                        'minLength' => 1,
+                        'maxLength' => 16,
+                    ),
+                ),
+                'AutoScalingGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+                'ShouldDecrementDesiredCapacity' => array(
+                    'required' => true,
+                    'type' => 'boolean',
+                    'format' => 'boolean-string',
+                    'location' => 'aws.query',
+                ),
+            ),
+        ),
         'ExecutePolicy' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -1358,6 +1578,112 @@ return array (
                 array(
                     'reason' => 'You cannot delete an Auto Scaling group while there are scaling activities in progress for that group.',
                     'class' => 'ScalingActivityInProgressException',
+                ),
+            ),
+        ),
+        'ExitStandby' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'ExitStandbyAnswer',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'ExitStandby',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2011-01-01',
+                ),
+                'InstanceIds' => array(
+                    'type' => 'array',
+                    'location' => 'aws.query',
+                    'sentAs' => 'InstanceIds.member',
+                    'items' => array(
+                        'name' => 'XmlStringMaxLen16',
+                        'type' => 'string',
+                        'minLength' => 1,
+                        'maxLength' => 16,
+                    ),
+                ),
+                'AutoScalingGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+            ),
+        ),
+        'PutLifecycleHook' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'PutLifecycleHook',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2011-01-01',
+                ),
+                'LifecycleHookName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 255,
+                ),
+                'AutoScalingGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+                'LifecycleTransition' => array(
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                ),
+                'RoleARN' => array(
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+                'NotificationTargetARN' => array(
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+                'NotificationMetadata' => array(
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1023,
+                ),
+                'HeartbeatTimeout' => array(
+                    'type' => 'numeric',
+                    'location' => 'aws.query',
+                ),
+                'DefaultResult' => array(
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'The quota for capacity groups or launch configurations for this customer has already been reached.',
+                    'class' => 'LimitExceededException',
                 ),
             ),
         ),
@@ -1556,6 +1882,46 @@ return array (
                 array(
                     'reason' => 'The quota for capacity groups or launch configurations for this customer has already been reached.',
                     'class' => 'LimitExceededException',
+                ),
+            ),
+        ),
+        'RecordLifecycleActionHeartbeat' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'RecordLifecycleActionHeartbeat',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2011-01-01',
+                ),
+                'LifecycleHookName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 255,
+                ),
+                'AutoScalingGroupName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 1,
+                    'maxLength' => 1600,
+                ),
+                'LifecycleActionToken' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 36,
+                    'maxLength' => 36,
                 ),
             ),
         ),
@@ -2228,6 +2594,65 @@ return array (
                 ),
             ),
         ),
+        'DescribeLifecycleHookTypesAnswer' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'LifecycleHookTypes' => array(
+                    'type' => 'array',
+                    'location' => 'xml',
+                    'items' => array(
+                        'name' => 'XmlStringMaxLen255',
+                        'type' => 'string',
+                        'sentAs' => 'member',
+                    ),
+                ),
+            ),
+        ),
+        'DescribeLifecycleHooksAnswer' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'LifecycleHooks' => array(
+                    'type' => 'array',
+                    'location' => 'xml',
+                    'items' => array(
+                        'name' => 'LifecycleHook',
+                        'type' => 'object',
+                        'sentAs' => 'member',
+                        'properties' => array(
+                            'LifecycleHookName' => array(
+                                'type' => 'string',
+                            ),
+                            'AutoScalingGroupName' => array(
+                                'type' => 'string',
+                            ),
+                            'LifecycleTransition' => array(
+                                'type' => 'string',
+                            ),
+                            'NotificationTargetARN' => array(
+                                'type' => 'string',
+                            ),
+                            'RoleARN' => array(
+                                'type' => 'string',
+                            ),
+                            'NotificationMetadata' => array(
+                                'type' => 'string',
+                            ),
+                            'HeartbeatTimeout' => array(
+                                'type' => 'numeric',
+                            ),
+                            'GlobalTimeout' => array(
+                                'type' => 'numeric',
+                            ),
+                            'DefaultResult' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+        ),
         'DescribeMetricCollectionTypesAnswer' => array(
             'type' => 'object',
             'additionalProperties' => true,
@@ -2519,6 +2944,147 @@ return array (
                         'name' => 'XmlStringMaxLen1600',
                         'type' => 'string',
                         'sentAs' => 'member',
+                    ),
+                ),
+            ),
+        ),
+        'DetachInstancesAnswer' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'Activities' => array(
+                    'type' => 'array',
+                    'location' => 'xml',
+                    'items' => array(
+                        'name' => 'Activity',
+                        'type' => 'object',
+                        'sentAs' => 'member',
+                        'properties' => array(
+                            'ActivityId' => array(
+                                'type' => 'string',
+                            ),
+                            'AutoScalingGroupName' => array(
+                                'type' => 'string',
+                            ),
+                            'Description' => array(
+                                'type' => 'string',
+                            ),
+                            'Cause' => array(
+                                'type' => 'string',
+                            ),
+                            'StartTime' => array(
+                                'type' => 'string',
+                            ),
+                            'EndTime' => array(
+                                'type' => 'string',
+                            ),
+                            'StatusCode' => array(
+                                'type' => 'string',
+                            ),
+                            'StatusMessage' => array(
+                                'type' => 'string',
+                            ),
+                            'Progress' => array(
+                                'type' => 'numeric',
+                            ),
+                            'Details' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+        ),
+        'EnterStandbyAnswer' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'Activities' => array(
+                    'type' => 'array',
+                    'location' => 'xml',
+                    'items' => array(
+                        'name' => 'Activity',
+                        'type' => 'object',
+                        'sentAs' => 'member',
+                        'properties' => array(
+                            'ActivityId' => array(
+                                'type' => 'string',
+                            ),
+                            'AutoScalingGroupName' => array(
+                                'type' => 'string',
+                            ),
+                            'Description' => array(
+                                'type' => 'string',
+                            ),
+                            'Cause' => array(
+                                'type' => 'string',
+                            ),
+                            'StartTime' => array(
+                                'type' => 'string',
+                            ),
+                            'EndTime' => array(
+                                'type' => 'string',
+                            ),
+                            'StatusCode' => array(
+                                'type' => 'string',
+                            ),
+                            'StatusMessage' => array(
+                                'type' => 'string',
+                            ),
+                            'Progress' => array(
+                                'type' => 'numeric',
+                            ),
+                            'Details' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+        ),
+        'ExitStandbyAnswer' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'Activities' => array(
+                    'type' => 'array',
+                    'location' => 'xml',
+                    'items' => array(
+                        'name' => 'Activity',
+                        'type' => 'object',
+                        'sentAs' => 'member',
+                        'properties' => array(
+                            'ActivityId' => array(
+                                'type' => 'string',
+                            ),
+                            'AutoScalingGroupName' => array(
+                                'type' => 'string',
+                            ),
+                            'Description' => array(
+                                'type' => 'string',
+                            ),
+                            'Cause' => array(
+                                'type' => 'string',
+                            ),
+                            'StartTime' => array(
+                                'type' => 'string',
+                            ),
+                            'EndTime' => array(
+                                'type' => 'string',
+                            ),
+                            'StatusCode' => array(
+                                'type' => 'string',
+                            ),
+                            'StatusMessage' => array(
+                                'type' => 'string',
+                            ),
+                            'Progress' => array(
+                                'type' => 'numeric',
+                            ),
+                            'Details' => array(
+                                'type' => 'string',
+                            ),
+                        ),
                     ),
                 ),
             ),